# Conflicts:
#	app/Http/Controllers/SprintController.php
#	resources/views/web/sections/profile/show.blade.php

+++ app/Http/Controllers/TaskController.php
@@ -61,7 +61,13 @@ class TaskController extends Controller
         // Authorization checks
         $userId = Auth::id();
 
-        if (in_array($newState, ['IN_PROGRESS', 'DONE', 'ACCEPTED']) && $task->assigned_to !== $userId) {
+        $project = Project::findOrFail($task->project_id);
+
+        if ($userId !== $project->product_owner_id && in_array($newState, ['IN_PROGRESS', 'DONE']) && $task->assigned_to !== $userId && $task->state === 'ACCEPTED') {
+            return response()->json(['status' => 'error', 'message' => 'You are not authorized to modify this task.'], 403);
+        }
+
+        if ($userId !== $project->product_owner_id && $task->state === 'ACCEPTED') {
             return response()->json(['status' => 'error', 'message' => 'You are not authorized to modify this task.'], 403);
         }
 
@@ -88,15 +94,29 @@ class TaskController extends Controller
     public function createNew(Request $request)
     {
         $project = Project::where('slug', $request->slug)->firstOrFail();
-        $task = new Task();
-        $task->project_id = $project->id;
-        $task->title = $request->title;
-        $task->description = $request->description;
-        $task->value = $request->value;
-        $task->effort = $request->effort;
-        $task->save();
 
-        return redirect()->route('projects.backlog', $project->slug)->with('success', 'Task created successfully!');
+        $validated = $request->validate([
+            'title' => 'required|string|max:200',
+            'description' => 'nullable|string|max:2000',
+            'effort' => 'required',
+            'value' => 'required',
+        ]);
+
+        $task = new Task([
+            'project_id' => $project->id,
+            'title' => $validated['title'],
+            'description' => $validated['description'] ?? 'No description',
+            'value' => $validated['value'],
+            'effort' => $validated['effort'],
+        ]);
+
+        if ($this->authorize('create', $task)){
+            $task->save();
+            return redirect()->route('projects.backlog', $project->slug)->with('success', 'Task created successfully!');
+        }
+        else{
+            return redirect()->back()->with('error', 'You are not authorized to create this task.');
+        }
     }
 
     public function showEdit($slug, $task_id){
@@ -105,24 +125,33 @@ class TaskController extends Controller
     }
 
     public function editTask(Request $request, $slug, $taskId){
-        $task = Task::findOrFail($taskId);
-        $task->title = $request->title;
-        $task->description = $request->description;
-        $task->value = $request->value;
-        $task->effort = $request->effort;
-        $task->save();
+        $validated = $request->validate([
+            'title' => 'required|string|max:200',
+            'description' => 'nullable|string|max:2000',
+            'effort' => 'required',
+            'value' => 'required',
+        ]);
 
-        return redirect()->route('task.show', $task)->with('success', 'Task updated successfully!');
+        $task = Task::findOrFail($taskId);
+        $task->title =  $validated['title'];
+        $task->description = $validated['description'] ?? 'No description';
+        $task->value = $validated['value'];
+        $task->effort = $validated['effort'];
+
+        if ($this->authorize('update', $task)){
+            $task->save();
+            return redirect()->route('task.show', $task)->with('success', 'Task updated successfully!');
+        }
+        else {
+            return redirect()->back()->with('error', 'You are not authorized to update this task.');
+        }
     }
 
     public function show($id)
     {
         $task = Task::where('id', $id)->firstOrFail();
-
         $sprint = $task->sprint;
-
         $project = $task->project;
-
         $comments = $task->comments()->orderBy('created_at', 'asc')->get();
 
         return view('web.sections.task.show', ['task' => $task, 'sprint' => $sprint, 'project' => $project, 'comments' => $comments,]);
@@ -133,9 +162,12 @@ class TaskController extends Controller
         $project = Project::where('slug', $slug)->firstOrFail();
         $task = Task::where('id', $id)->where('project_id', $project->id)->firstOrFail();
 
-        $task->delete();
-
-        return redirect()->route('projects.backlog', ['slug' => $slug])
-            ->with('success', 'Task deleted successfully.');
+        if ($this->authorize('delete', $task)){
+            $task->delete();
+            return redirect()->route('projects.backlog', ['slug' => $slug])->with('success', 'Task deleted successfully.');
+        }
+        else{
+            return redirect()->back()->with('error', 'You are not authorized to delete this task.');
+        }
     }
 }
